fix the fallback boot entry
my system forgets the boot entry if that drive is removed, and I found out the failback efi entry is not working (just take me to the UEFI settings). Copying everything under EFI/syslinux over seems to fixed the problem.

I'm very new to ChimeraOS so I'm not sure if this is the only place that needs updating (like updates or whatever). and really I think it might be a good idea for ChimeraOS to just use the default location for everything since its the only os on the drive anyway.