Allow custom boot paramenters in image
Since most arguments are tied to the kernel version in the image, we
should be able to accomodate for that. For example an argument like
`split_lock_detect=off` only makes sense for gaming images on kernel
version 6.1 and after.

This will look inside the image for additional arguments to put into the
bootloader configuration.

Arguments need to be written in a special file
(`/usr/lib/frzr.d/bootconfig.conf`) in a single line separated by spaces
or one parameter per line. The contents of that file will be `cat`-ed to
the current bootloader arguments.