tests: Each channel requested should be respected
Each channel should be respected as the user requested. Even if that
means a downgrade from current installed version.

That means if the user requested `unstable` it should check for the
latest unstable image regardless of stable one. The same goes for
`testing`. Each channel will prioritize it's own latest image avobe the
stable one.